+++ src/lexer.cpp
@@ -6,7 +6,6 @@
 *********************************************************************************/
 
 
-#include <XPD.h>
 #include "error.h"
 #include "lexer.h"
 #include "utility.h"
@@ -26,7 +25,7 @@
  */
 Lexer::Lexer(char ** input, lexed_command * output) {
     // transfer input string into object
-    for (uint16_t i = 0; i < MAX_INPUT_LEN; i++) {
+    for (int i = 0; i < MAX_INPUT_LEN; i++) {
         line[i] = *(*input + i);
     }
     // get the number of non-null characters in the command
@@ -45,7 +44,7 @@ void Lexer::scan_next_token() {
     // to store string and number literals as well as identifiers
     char str_lit[MAX_LIT_LEN] = "";
     char * str_lit_ptr = (char *) str_lit;
-    uint16_t num_lit = 0;
+    int num_lit = 0;
     char identifier[MAX_IDENTIFIER_LEN] = "";
     char * identifier_ptr = (char *) identifier;
 
@@ -260,7 +259,7 @@ void Lexer::add_token(lexemes token) {
  * \param [in] str_lit The string literal to be added to the list.
  */
 void Lexer::add_str_lit(char * str_lit) {
-    for (uint16_t i = 0; i < MAX_LIT_LEN; i++) {
+    for (int i = 0; i < MAX_LIT_LEN; i++) {
         command_info -> str_lits[command_info -> str_lit_count][i] = *(str_lit + i);
     }
     command_info -> str_lit_count++;
@@ -271,7 +270,7 @@ void Lexer::add_str_lit(char * str_lit) {
  * \brief Adds a number literal to the list for the instruction.
  * \param [in] num_lit The number literal to be added to the list.
  */
-void Lexer::add_num_lit(uint16_t num_lit) {
+void Lexer::add_num_lit(int num_lit) {
     command_info -> num_lits[command_info -> num_lit_count] = num_lit;
     command_info -> num_lit_count++;
 }
@@ -282,7 +281,7 @@ void Lexer::add_num_lit(uint16_t num_lit) {
  * \param [in] identifier The identifier to be added to the list.
  */
 void Lexer::add_identifier(char * identifier) {
-    for (uint16_t i = 0; i < MAX_LIT_LEN; i++) {
+    for (int i = 0; i < MAX_LIT_LEN; i++) {
         command_info -> identifiers[command_info -> identifier_count][i] = *(identifier + i);
     }
     command_info -> identifier_count++;
@@ -322,7 +321,7 @@ bool Lexer::next_matches(char character) {
  */
 void Lexer::match_string(char terminator, char ** output_ptr) {
     // make an index to track the literal's value string
-    uint16_t i = 0;
+    int i = 0;
     // start at character after the leading delimiter
     current++;
     // add each character to the literla that is not its terminator
@@ -344,12 +343,12 @@ void Lexer::match_string(char terminator, char ** output_ptr) {
  * \brief Recovers the number literal in the instruction.
  * \param [in] output_ptr Pointer to where to store the literal value.
  */
-void Lexer::match_number(uint16_t * output_ptr) {
+void Lexer::match_number(int * output_ptr) {
     // track the number string to be parsed
     char num_str[MAX_LIT_LEN] = "";
     char * num_str_ptr = (char *) num_str;
     // add all numerical characters to a cumulative string
-    uint16_t i = 0;
+    int i = 0;
     while (isdigit(line[current]) && (i < MAX_LIT_LEN)) {
         *(num_str_ptr + i) = line[current];
         i++;
@@ -367,7 +366,7 @@ void Lexer::match_number(uint16_t * output_ptr) {
  * \param [in] output_ptr Pointer to where to store the identifier name.
  */
 void Lexer::match_identifier(char ** output_ptr) {
-    uint16_t i = 0;
+    int i = 0;
     // add each character to the identifier until no more eligible characters
     while (isalphanumeric(line[current]) && (i < MAX_IDENTIFIER_LEN)) {
         *(*output_ptr + i) = line[current];
@@ -685,7 +684,7 @@ bool Lexer::has_error() {
  * \brief Scans the input string and returns a list of its tokens.
  * \return 0 if execution succeeded; non-zero value if an error occurred.
  */
-uint16_t Lexer::scan_input() {
+int Lexer::scan_input() {
     // keep on reading next character until command is over
     while (!(end_reached())) {
         scan_next_token();
